add ToDouble

+++ task21/Program.cs
@@ -4,13 +4,13 @@
 // A (7,-5, 0); B (1,-1,9) -> 11.53
 
 Console.WriteLine("Введите координаты для точки А: ");
-double ax = Convert.ToInt32(Console.ReadLine());
-double ay = Convert.ToInt32(Console.ReadLine());
-double az = Convert.ToInt32(Console.ReadLine());
+double ax = Convert.ToDouble(Console.ReadLine());
+double ay = Convert.ToDouble(Console.ReadLine());
+double az = Convert.ToDouble(Console.ReadLine());
 Console.WriteLine("Введите координаты для точки В: ");
-double bx = Convert.ToInt32(Console.ReadLine());
-double by = Convert.ToInt32(Console.ReadLine());
-double bz = Convert.ToInt32(Console.ReadLine());
+double bx = Convert.ToDouble(Console.ReadLine());
+double by = Convert.ToDouble(Console.ReadLine());
+double bz = Convert.ToDouble(Console.ReadLine());
 
 double distans = Math.Sqrt(Math.Pow((ax-bx), 2) + Math.Pow((ay-by), 2) + Math.Pow((az-bz), 2));
